Refactor SyncButton component: migrate from container to functional component and update imports
