+++ src/components/Communicator/CommunicatorToolbar/SyncButton/SyncButton.js
@@ -1,6 +1,8 @@
 import React, { useState, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
+import { connect } from 'react-redux';
+import { injectIntl } from 'react-intl';
 import Button from '@material-ui/core/Button';
 import IconButton from '@material-ui/core/IconButton';
 import CircularProgress from '@material-ui/core/CircularProgress';
@@ -11,6 +13,8 @@ import SyncProblemIcon from '@material-ui/icons/SyncProblem';
 import green from '@material-ui/core/colors/green';
 import amber from '@material-ui/core/colors/amber';
 import { SYNC_STATUS } from '../../../Board/Board.constants';
+import { hasPendingSyncBoards } from '../../../Board/Board.selectors';
+import { getApiMyBoards } from '../../../Board/Board.actions';
 import messages from './SyncButton.messages';
 import { SYNCED_DISPLAY_DURATION, DISPLAY_STATE } from './SyncButton.constants';
 import './SyncButton.css';
@@ -182,4 +186,19 @@ SyncButton.defaultProps = {
   onSyncClick: () => {}
 };
 
-export default SyncButton;
+const mapStateToProps = (state, ownProps) => ({
+  isOnline: state.app.isConnected,
+  isSyncing: state.board.isSyncing,
+  isFetching: state.board.isFetching,
+  isSaving: ownProps.isSaving || false,
+  hasPendingBoards: hasPendingSyncBoards(state)
+});
+
+const mapDispatchToProps = {
+  onSyncClick: getApiMyBoards
+};
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(injectIntl(SyncButton));
